Enable building on AIX
The AIX operating system by IBM does not provide the getifaddrs function
and requires a custom linker script. This commit allows for a basic GCC
build on AIX 7.2 using the cmake build system. Building with plugin
support does not seem to be possible, yet.

The commit was tested using the command line

CC=gcc cmake ../mosquitto/ -DWITH_PLUGINS=no -DDOCUMENTATION=no

The library libmosquitto.a as well as the mosquitto_sub and
mosquitto_pub executables were tested and appear to be fully functional.